requirements.txt
alembic==1.2.1
asn1crypto==1.1.0
blinker==1.4
cffi==1.13.0
Click==7.0
cryptography==2.7
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
